+++ lib/essentials/navigation/presentation/view/macos_app_shell.dart
@@ -30,7 +30,6 @@ import '../../../onboarding/feature_level_providers.dart'
     show onboardingGateProvider, onboardingOperationReconciliationProvider;
 import '../../../onboarding/presentation/advanced_start_fresh_overlay.dart';
 import '../../../onboarding/presentation/onboarding_overlay.dart';
-import '../../../onboarding/presentation/onboarding_presence_host.dart';
 import '../../../sidebar/application/sidebar_flow_state_provider.dart';
 import '../../application/app_shell_actions_provider.dart';
 import '../../application/panel_widget_providers.dart';
@@ -139,11 +138,6 @@ class _MacosAppShellState extends ConsumerState<MacosAppShell> {
       OnboardingStatus.reimporting ||
       OnboardingStatus.reimportBuildingGraph ||
       OnboardingStatus.reimportComplete => true,
-      _ => false,
-    };
-    final showRequiredSourcesPresence = switch (onboardingStatus) {
-      OnboardingStatus.awaitingFda ||
-      OnboardingStatus.awaitingUserAction => true,
       _ => false,
     };
     final activeMode = ref.watch(activeSidebarModeProvider);
@@ -516,7 +510,6 @@ class _MacosAppShellState extends ConsumerState<MacosAppShell> {
     return Stack(
       children: <Widget>[
         window,
-        if (showRequiredSourcesPresence) const OnboardingPresenceHost(),
         if (showOnboardingOverlay) const OnboardingOverlay(),
         const AdvancedStartFreshOverlayHost(),
       ],